# gift_converter/openapi.json
{"openapi":"3.1.0","info":{"title":"GIFT Converter API","version":"1.0.0"},"paths":{"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/gift-to-json":{"post":{"summary":"Gift To Json","operationId":"gift_to_json_gift_to_json_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_gift_to_json_gift_to_json_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConvertResponseSchema"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/json-to-gift":{"post":{"summary":"Json To Gift","operationId":"json_to_gift_json_to_gift_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_json_to_gift_json_to_gift_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_gift_to_json_gift_to_json_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"}},"type":"object","required":["file"],"title":"Body_gift_to_json_gift_to_json_post"},"Body_json_to_gift_json_to_gift_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"}},"type":"object","required":["file"],"title":"Body_json_to_gift_json_to_gift_post"},"ChoiceSchema":{"properties":{"text":{"type":"string","title":"Text"},"is_correct":{"type":"boolean","title":"Is Correct"},"weight":{"type":"number","title":"Weight"},"feedback":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Feedback"}},"type":"object","required":["text","is_correct","weight"],"title":"ChoiceSchema"},"ConvertResponseSchema":{"properties":{"created_at":{"type":"string","title":"Created At"},"total_questions":{"type":"integer","title":"Total Questions"},"questions":{"items":{"$ref":"#/components/schemas/QuestionSchema"},"type":"array","title":"Questions"}},"type":"object","required":["created_at","total_questions","questions"],"title":"ConvertResponseSchema"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MatchingPairSchema":{"properties":{"term":{"type":"string","title":"Term"},"definition":{"type":"string","title":"Definition"}},"type":"object","required":["term","definition"],"title":"MatchingPairSchema"},"QuestionSchema":{"properties":{"question_type":{"type":"string","title":"Question Type"},"text":{"type":"string","title":"Text"},"title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Title"},"choices":{"anyOf":[{"items":{"$ref":"#/components/schemas/ChoiceSchema"},"type":"array"},{"type":"null"}],"title":"Choices"},"short_answers":{"anyOf":[{"items":{"$ref":"#/components/schemas/ShortAnswerSchema"},"type":"array"},{"type":"null"}],"title":"Short Answers"},"matching_pairs":{"anyOf":[{"items":{"$ref":"#/components/schemas/MatchingPairSchema"},"type":"array"},{"type":"null"}],"title":"Matching Pairs"}},"type":"object","required":["question_type","text"],"title":"QuestionSchema"},"ShortAnswerSchema":{"properties":{"text":{"type":"string","title":"Text"},"weight":{"type":"number","title":"Weight"}},"type":"object","required":["text","weight"],"title":"ShortAnswerSchema"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}